+++ app/hooks/useRoleStore.ts
@@ -1,20 +1,19 @@
 import { coffeApi } from '@/services';
-import { useAlertStore, useErrorStore } from '.';
-import { InitBaseResponse, type BaseResponse, type RoleModel } from '@/models';
+import { useAlertStore, useErrorStore, usePermissionStore } from '.';
+import { InitBaseResponse, TypeAction, TypeSubject, type BaseResponse, type RoleModel } from '@/models';
 import type { RoleRequest } from '@/models/request/role.request';
 import { useState } from 'react';
 
 export const useRoleStore = () => {
   const [dataRole, setDataRole] = useState<BaseResponse<RoleModel>>(InitBaseResponse);
-
   const { handleError } = useErrorStore();
+  const { requirePermission } = usePermissionStore();
   const { showSuccess, showWarning, showError } = useAlertStore();
   const baseUrl = 'role';
 
-
-
   const getRoles = async (page: number = 1, limit: number = 10, keys: string = '') => {
     try {
+      requirePermission(TypeAction.read, TypeSubject.role);
       const res = await coffeApi.get(`/${baseUrl}?page=${page}&limit=${limit}&keys=${keys}`);
       const { data, meta } = res.data;
       console.log(res.data);
@@ -30,6 +29,7 @@ export const useRoleStore = () => {
 
   const createRole = async (body: RoleRequest) => {
     try {
+      requirePermission(TypeAction.create, TypeSubject.role);
       const { data } = await coffeApi.post(`/${baseUrl}`, body);
       console.log(data);
       getRoles();
@@ -41,6 +41,7 @@ export const useRoleStore = () => {
 
   const updateRole = async (id: string, body: RoleRequest) => {
     try {
+      requirePermission(TypeAction.update, TypeSubject.role);
       const { data } = await coffeApi.patch(`/${baseUrl}/${id}`, body);
       console.log(data);
       getRoles();
@@ -52,6 +53,7 @@ export const useRoleStore = () => {
 
   const deleteRole = async (id: string) => {
     try {
+      requirePermission(TypeAction.delete, TypeSubject.role);
       const result = await showWarning();
       if (result.isConfirmed) {
         await coffeApi.delete(`/${baseUrl}/${id}`);
