+++ app/hooks/useRoomStore.ts
@@ -1,16 +1,18 @@
 import { coffeApi } from '@/services';
-import { useAlertStore, useErrorStore } from '.';
-import { InitBaseResponse, type BaseResponse, type RoomModel, type RoomRequest } from '@/models';
+import { useAlertStore, useErrorStore, usePermissionStore } from '.';
+import { InitBaseResponse, TypeAction, TypeSubject, type BaseResponse, type RoomModel, type RoomRequest } from '@/models';
 import { useState } from 'react';
 
 export const useRoomStore = () => {
   const [dataRoom, setDataRoom] = useState<BaseResponse<RoomModel>>(InitBaseResponse);
   const { handleError } = useErrorStore();
+  const { requirePermission } = usePermissionStore();
   const { showSuccess, showWarning, showError } = useAlertStore();
   const baseUrl = 'room';
 
   const getRooms = async (page: number = 1, limit: number = 10, keys: string = '') => {
     try {
+      requirePermission(TypeAction.read, TypeSubject.room);
       const res = await coffeApi.get(`/${baseUrl}?page=${page}&limit=${limit}&keys=${keys}`);
       const { data, meta } = res.data;
       console.log(res.data);
@@ -25,6 +27,7 @@ export const useRoomStore = () => {
   };
   const createRoom = async (body: RoomRequest) => {
     try {
+      requirePermission(TypeAction.create, TypeSubject.room);
       const { data } = await coffeApi.post(`/${baseUrl}/`, body);
       console.log(data);
       getRooms();
@@ -35,6 +38,7 @@ export const useRoomStore = () => {
   };
   const updateRoom = async (id: string, body: RoomRequest) => {
     try {
+      requirePermission(TypeAction.update, TypeSubject.room);
       const { data } = await coffeApi.patch(`/${baseUrl}/${id}`, body);
       console.log(data);
       getRooms();
@@ -45,6 +49,7 @@ export const useRoomStore = () => {
   };
   const deleteRoom = async (id: string) => {
     try {
+      requirePermission(TypeAction.delete, TypeSubject.room);
       const result = await showWarning();
       if (result.isConfirmed) {
         await coffeApi.delete(`/${baseUrl}/${id}`);
