+++ app/hooks/useSpecialtyStore.ts
@@ -1,18 +1,18 @@
-import { useDispatch } from 'react-redux';
 import { coffeApi } from '@/services';
-import { useAlertStore, useErrorStore } from '.';
-import { InitBaseResponse, type BaseResponse, type SpecialtyModel, type SpecialtyRequest } from '@/models';
+import { useAlertStore, useErrorStore, usePermissionStore } from '.';
+import { InitBaseResponse, TypeAction, TypeSubject, type BaseResponse, type SpecialtyModel, type SpecialtyRequest } from '@/models';
 import { useState } from 'react';
 
 export const useSpecialtyStore = () => {
   const [dataSpecialty, setDataSpecialty] = useState<BaseResponse<SpecialtyModel>>(InitBaseResponse);
-  const dispatch = useDispatch();
   const { handleError } = useErrorStore();
+  const { requirePermission } = usePermissionStore();
   const { showSuccess, showWarning, showError } = useAlertStore();
   const baseUrl = 'specialty';
 
   const getSpecialties = async (page: number = 1, limit: number = 10, keys: string = '') => {
     try {
+      requirePermission(TypeAction.read, TypeSubject.specialty);
       const res = await coffeApi.get(`/${baseUrl}?page=${page}&limit=${limit}&keys=${keys}`);
       const { data, meta } = res.data;
       console.log(res.data);
@@ -28,6 +28,7 @@ export const useSpecialtyStore = () => {
 
   const getSpecialtiesByBranch = async (branchId: string, page: number = 1, limit: number = 10, keys: string = '') => {
     try {
+      requirePermission(TypeAction.read, TypeSubject.specialty);
       const res = await coffeApi.get(`/${baseUrl}/branch/${branchId}?page=${page}&limit=${limit}&keys=${keys}`);
       const { data, meta } = res.data;
       console.log(res.data);
@@ -43,6 +44,7 @@ export const useSpecialtyStore = () => {
 
   const createSpecialty = async (body: SpecialtyRequest) => {
     try {
+      requirePermission(TypeAction.create, TypeSubject.specialty);
       const { data } = await coffeApi.post(`${baseUrl}`, body);
       console.log(data);
       getSpecialties();
@@ -53,6 +55,7 @@ export const useSpecialtyStore = () => {
   };
   const updateSpecialty = async (id: string, body: SpecialtyRequest) => {
     try {
+      requirePermission(TypeAction.update, TypeSubject.specialty);
       const { data } = await coffeApi.patch(`/${baseUrl}/${id}`, body);
       console.log(data);
       getSpecialties();
@@ -63,6 +66,7 @@ export const useSpecialtyStore = () => {
   };
   const deleteSpecialty = async (id: string) => {
     try {
+      requirePermission(TypeAction.delete, TypeSubject.specialty);
       const result = await showWarning();
       if (result.isConfirmed) {
         await coffeApi.delete(`/${baseUrl}/${id}`);
